:bug: Make circle constant dependent on sign of ang2.
Previous to the commit where the spencermortensen value was added as an
improvement to the circle approximation, the constant for circle
approximation was actually generated dependent on the value of one of the
angles passed as input. This allowed the preservation of the sign.
Unfortunately, by hard-coding a constant, we lost this flexibility. This
once again makes the value dependent on the sign of the input angle.

Fixes #7.